Corrige video

+++ src/Pages/Profile.jsx
@@ -1,19 +1,48 @@
 import React from 'react';
 import Header from '../Components/Header';
-import Footer from '../Components/Footer';
 import ButtonCard from '../Components/ButtonCard';
+import Footer from '../Components/Footer';
 
 function Profile() {
   const getUser = localStorage.getItem('user');
   const user = JSON.parse(getUser);
 
-  const onClickLogout = () => {
-    localStorage.clear();
-  };
+  if (getUser === null) {
+    return (
+      <>
+        <Header title="Perfil" />
+        <h3
+          data-testid="profile-email"
+        >
+          email@email.com
+        </h3>
+        <ButtonCard
+          page="/receitas-feitas"
+          testId="profile-done-btn"
+          buttonText="Receitas Feitas"
+        />
+        <ButtonCard
+          page="/receitas-favoritas"
+          testId="profile-favorite-btn"
+          buttonText="Receitas Favoritas"
+        />
+        <ButtonCard
+          page="/"
+          testId="profile-logout-btn"
+          onClick={ () => {
+            localStorage.removeItem('cocktailsToken', 'mealsToken');
+            localStorage.clear();
+          } }
+          buttonText="Sair"
+        />
+        <Footer />
+      </>
+    );
+  }
 
   return (
     <>
-      <Header title="Perfil" loading={ false } />
+      <Header title="Perfil" />
       <h3
         data-testid="profile-email"
       >
@@ -32,7 +61,10 @@ function Profile() {
       <ButtonCard
         page="/"
         testId="profile-logout-btn"
-        onClick={ onClickLogout }
+        onClick={ () => {
+          localStorage.removeItem('cocktailsToken', 'mealsToken');
+          localStorage.clear();
+        } }
         buttonText="Sair"
       />
       <Footer />
